refactor: extract context usage indicator from EngineSelect
