+++ libs/shared/src/components/engine/index.ts
@@ -1,2 +1,3 @@
+export * from "./context-usage-indicator";
 export * from "./engine-select";
 export * from "./new-engine-input";
